Refactoring. Spring pur.
