+++ pom.xml
@@ -18,7 +18,6 @@
 		<jstl.version>1.2</jstl.version>
 		<servlet.api.version>3.0.1</servlet.api.version>
 		<jsp.api.version>2.2</jsp.api.version>
-		<inject.version>1</inject.version>
 		<spring.version>3.2.3.RELEASE</spring.version>
 		<hibernate.version>4.2.1.Final</hibernate.version>
 		<c3p0.version>0.9.2.1</c3p0.version>
@@ -93,14 +92,6 @@
 			<artifactId>jsp-api</artifactId>
 			<version>${jsp.api.version}</version>
 			<scope>provided</scope>
-		</dependency>
-
-		<!-- JSR-330 -->
-		<dependency>
-			<groupId>javax.inject</groupId>
-			<artifactId>javax.inject</artifactId>
-			<version>${inject.version}</version>
-			<scope>compile</scope>
 		</dependency>
 
 		<!-- Spring framework -->
