+++ src/main/java/com/jeetemplates/app/common/dto/GenericDTO.java
@@ -1,0 +1,20 @@
+/**
+ * 
+ */
+package com.jeetemplates.app.common.dto;
+
+import java.io.Serializable;
+
+/**
+ * Common business dto
+ * 
+ * @author jeetemplates
+ */
+public abstract class GenericDTO implements Serializable {
+
+	/**
+	 * Serial UID.
+	 */
+	private static final long serialVersionUID = 3886269150595934285L;
+
+}
