--- src/main/java/com/jeetemplates/common/dto/GenericDTO.java
@@ -1,20 +1,0 @@
-/**
- * 
- */
-package com.jeetemplates.common.dto;
-
-import java.io.Serializable;
-
-/**
- * Common business dto
- * 
- * @author jeetemplates
- */
-public abstract class GenericDTO implements Serializable {
-
-	/**
-	 * Serial UID.
-	 */
-	private static final long serialVersionUID = 3886269150595934285L;
-
-}
