--- src/main/java/com/jeetemplates/persistence/HelloWorldDao.java
@@ -1,16 +1,0 @@
-/**
- * 
- */
-package com.jeetemplates.persistence;
-
-import com.jeetemplates.common.persistence.BusinessDao;
-import com.jeetemplates.domain.model.HelloWorld;
-
-/**
- * DAO of HelloWorld
- * 
- * @author jeetemplates
- */
-public interface HelloWorldDao extends BusinessDao<HelloWorld> {
-
-}
