+++ src/main/resources/spring/spring-beans.xml
@@ -2,11 +2,11 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:context="http://www.springframework.org/schema/context"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans
-	http://www.springframework.org/schema/beans/spring-beans-3.1.xsd
+	http://www.springframework.org/schema/beans/spring-beans.xsd
 	http://www.springframework.org/schema/context
-	http://www.springframework.org/schema/context/spring-context-3.1.xsd">
+	http://www.springframework.org/schema/context/spring-context.xsd">
 
-	<!-- Scans the classpath of this application for @Named to deploy as beans -->
-	<context:component-scan base-package="com.jeetemplates" />
-
+	<!-- Scans the classpath of this application to deploy as beans -->
+	<context:component-scan base-package="com.jeetemplates.app" annotation-config="true"/>
+	
 </beans>
