+++ src/main/webapp/pages/template/layout.xhtml
@@ -7,7 +7,7 @@
 	  xmlns:ui="http://java.sun.com/jsf/facelets">
 
 
-	<f:view locale="#{headerController.locale}">
+	<f:view locale="#{localeBean.locale}">
 		<h:head>
 			<h:outputStylesheet name="common-style.css" library="css" />
 		</h:head>
